Port Slow Stairs Movement From FRLG
This should've been included in the PR that added the functionality to walk slow on stairs from FRLG.

In FRLG, walking on stairs uses a custom movement speed that is faster than Emerald's slow movement speed (which is the speed incorrectly used for walking on stairs in RHH and ghoul's sideways stairs branch).

In pokefirered, the movement labeled "slow" is only used when moving on stairs and the movement labeled "slower" is equivalent to "slow" in pokeemerald. This is likely where the confusion stems from.

This commit ports the stairs speed from pokefirered and additionally introduces movement commands for use in `applymovement` for this speed. Followers now also use the stairs speed where appropriate.