Port FRLG Script Walkaway
This probably should've been included with the PR that added the functionality to walk away from signposts sourced from FRLG, as this feature goes together with that.

In FRLG, if a script is about to end, the player can walk or turn away to exit the script and is not locked in place until they press the A or B button like in RSE. This commit ports that functionality.

FRLG includes functions to disable and reenable this feature, but reenabling (`EnableMsgBoxWalkaway`) is never actually used. Disabling (`DisableMsgBoxWalkaway`) was made into a special and is now used in the equivalent scripts to FRLG's usages.